+++ latestVersion.js
@@ -1,3 +1,3 @@
 // IMPORTANT: This are check by other version, change only when you want release!
 // IMPORTANT: Update also login.php#version, latestVersion.js, main.php
-var latestVersion = '1.9b';
+var latestVersion = '2.0.0b';