+++ main.php
@@ -17,6 +17,7 @@
 
     <link rel="stylesheet" href="//netdna.bootstrapcdn.com/font-awesome/4.2.0/css/font-awesome.min.css"/>
     <link rel="stylesheet" href="//ajax.googleapis.com/ajax/libs/jquerymobile/1.4.3/jquery.mobile.min.css"/>
+
     <link rel="stylesheet" type="text/css" href="addtohomescreen.css">
     <link rel="stylesheet" href="main.css"/>
 
@@ -26,11 +27,12 @@
     <script src="chart.min.js"></script>
     <script src="addtohomescreen.min.js"></script>
 </head>
+
 <body>
-<div data-role="page" id="main" class="type-interior">
+<div id="main" data-role="page">
 
     <div id="header" data-role="header" data-position="fixed">
-        <a href="#menu-panel" data-icon="home" class="ui-btn-left">Menu</a>
+        <a href="#menu-panel" class="ui-btn-left"><i class="fa fa-bars fa-fw"></i><span> Menu</span></a>
 
         <h1><i class="fa fa-home fa-fw"></i> Status</h1>
         <a href="#" data-rel="back" data-icon="arrow-l" class="ui-btn-right">Back</a>
@@ -42,12 +44,30 @@
     </div>
     <!-- /content -->
 
-    <div data-role="footer" role="contentinfo" class="ui-footer">
-        <a href="#" id="btScrollUp" data-role="button" data-icon="arrow-u" class="ui-btn-right">scroll up</a>
+    <div id="footer" data-role="footer">
+        <a id="btScrollUp" href="#" data-icon="arrow-u" class="ui-btn-right">scroll up</a>
 
         <p>&nbsp;<a href="#about" data-rel="popup" data-position-to="window" data-transition="pop">mobileMule</a> &copy; 2014</p>
     </div>
     <!-- /footer -->
+
+    <div data-role="panel" id="menu-panel" data-display="reveal" data-position-fixed="true">
+        <ul data-role="listview">
+            <li data-icon="false"><a href="#page-status" class="hash-link">Status <i class="fa fa-home pull-right fa-fw"></i></a></li>
+            <li data-icon="false"><a href="#page-downloads" class="hash-link">Downloads <i class="fa fa-download pull-right fa-fw"></i></a></li>
+            <li data-icon="false"><a href="#page-uploads" class="hash-link">Uploads <i class="fa fa-upload pull-right fa-fw"></i></a></li>
+            <li data-icon="false"><a href="#page-donate" class="hash-link">Search <i class="fa fa-search pull-right fa-fw"></i></a></li>
+            <li data-icon="false"><a href="#page-donate" class="hash-link">Settings <i class="fa fa-cogs pull-right fa-fw"></i></a></li>
+            <li data-icon="false"><a href="#page-donate" class="hash-link">MobileMule <i class="fa fa-cog pull-right fa-fw"></i></a></li>
+            <li data-icon="false"><a href="#page-servers" class="hash-link">Servers <i class="fa fa-database pull-right fa-fw"></i></a></li>
+            <li data-icon="false"><a href="#page-listed-stats" class="hash-link">Statistics <i class="fa fa-tachometer pull-right fa-fw"></i></a></li>
+            <li data-icon="false"><a href="#page-graph" class="hash-link">Graphs <i class="fa fa-bar-chart-o pull-right fa-fw"></i></a></li>
+            <li data-icon="false"><a href="#page-log" class="hash-link">Log <i class="fa fa-bars pull-right fa-fw"></i></a></li>
+            <li data-icon="false"><a href="#page-donate" class="hash-link">Add ed2k <i class="fa fa-plus pull-right fa-fw"></i></a></li>
+            <li data-icon="false"><a href="login.php" id="pnMenuLogout">Logout <i class="fa fa-unlock-alt pull-right fa-fw"></i></a></li>
+        </ul>
+    </div>
+    <!-- /panel menu -->
 
     <div data-role="popup" id="about">
         <div data-role="header">
@@ -58,10 +78,7 @@
         </div>
         <div role="main" class="ui-content">
             <!-- IMPORTANT: Update also login.php#version, latestVersion.js, main.php -->
-            <p><strong>mobileMule v2.0.0b</strong><br/>
-
-                <br/><i>Donation Package</i><br/>
-                Thanks for your Donation!<br/><br/>
+            <p><strong>mobileMule v2.1.0b</strong><br/>
 
                 coded by muttley &copy; copyright 2014<br/><br/>
                 <a href="https://raw.githubusercontent.com/elbowz/mobileMule/master/CHANGELOG.md" title="changelog">Changelog</a> &bull;
@@ -74,22 +91,6 @@
     <!-- /about -->
 </div>
 <!-- /page -->
-
-<div data-role="panel" id="menu-panel" data-display="reveal" data-position-fixed="true" data-theme="a">
-    <ul data-role="listview">
-        <li data-icon="false"><a href="#page-status" class="hash-link">Status <i class="fa fa-home pull-right fa-fw"></i></a></li>
-        <li data-icon="false"><a href="#page-downloads" class="hash-link">Downloads <i class="fa fa-download pull-right fa-fw"></i></a></li>
-        <li data-icon="false"><a href="#page-uploads" class="hash-link">Uploads <i class="fa fa-upload pull-right fa-fw"></i></a></li>
-        <li data-icon="false"><a href="#page-donate" class="hash-link">Search <i class="fa fa-search pull-right fa-fw"></i></a></li>
-        <li data-icon="false"><a href="#page-donate" class="hash-link">Settings <i class="fa fa-cogs pull-right fa-fw"></i></a></li>
-        <li data-icon="false"><a href="#page-servers" class="hash-link">Servers <i class="fa fa-database pull-right fa-fw"></i></a></li>
-        <li data-icon="false"><a href="#page-listed-stats" class="hash-link">Statistics <i class="fa fa-tachometer pull-right fa-fw"></i></a></li>
-        <li data-icon="false"><a href="#page-graph" class="hash-link">Graphs <i class="fa fa-bar-chart-o pull-right fa-fw"></i></a></li>
-        <li data-icon="false"><a href="#page-log" class="hash-link">Log <i class="fa fa-bars pull-right fa-fw"></i></a></li>
-        <li data-icon="false"><a href="#page-donate" class="hash-link">Add ed2k <i class="fa fa-plus pull-right fa-fw"></i></a></li>
-        <li data-icon="false"><a href="login.php" id="pnMenuLogout">Logout <i class="fa fa-unlock-alt pull-right fa-fw"></i></a></li>
-    </ul>
-</div>
 
 <script>
     addToHomescreen({ maxDisplayCount: 4 });