+++ CHANGELOG.md
@@ -8,8 +8,9 @@ CHANGELOG
  * Real-time Speed Chart on status page
  * Servers page now highlight the server connected
  * Add ed2k link now support multi-links
- * Added icons for Android and Apple devices
- * And more I don't remember
+ * Autologin (remember me)
+ * Added icons for Android and Apple devices when app is installed on home.
+ * And more and more...I don't remember
 
 ### 1.9b
  * Hide address bar in Chrome and Safari (Android and iOS) when user install Web Apps to the home screen (like native app!)